+++ src/main/java/ru/itpark/Main.java
@@ -3,9 +3,9 @@ package ru.itpark;
 public class Main {
     public static void main(String[] args) {
 
-        double purchaseBasic = 23_000;
-        double purchaseHight = 6_000;
-        double purchaseSpecial = 2_330;
+        int purchaseBasic = 23_000;
+        int purchaseHight = 6_000;
+        int purchaseSpecial = 2_330;
 
         CashbackService cashbackService = new CashbackService();
         int cashback = cashbackService.cashbackCalculate(purchaseBasic, purchaseHight, purchaseSpecial);
